app/requirements.txt
paho-mqtt==1.6.1
python-dotenv==1.0.0
boto3==1.28.32
psycopg2==2.9.7